scaleable product-search functions
executing view functions doesn't cost us any gas... but nevertheless
we can't be too cavalier about rediculiously inefficient functions.
since the contract might some day be required to handle hundreds of
thousands of products, it would be silly to require the search function
to search the entire list, for example to find the teensy number of
products offered by a certain vendor, or in a certain region.

the new search functions are getVendorProducts, getCategoryProducts,
and getRegionProducts. these functions only searche within the product
lists of a specific vendor, or a specific top-level category or top-
level region.

btw, categories and regions are now specified as 8 bits of top-level,
and 248 bits of lower level. the top level categories must be an exact
match with the search parameters; the lower levels categories are a bit
mask, and we only look for overlap between the product specification
and the search paremeter.