fix duplicate edited products bug
there are multiple event logs for products that are edited. so first
off we traverse the logs backwards, from mosr recent to oldest. that
way we encounter the most recent first. next we only process products
whose id's we hanv't already seen. also we only count each unique product
id once for the purposes of no. products processed in the doneCB.