mtUtil encrypt now returns msgNo
this is the right place to inc msgNo, because it is associated with
the encrypted message. now waitForTxid always returns {err, receipt},
and if there's no err then mtUtil.sentMsgCount is assigned from the
returned msgNo.

also now always clear waitingForTxid in the waitForTxid callback --
it is only cleared automatically in waitForTxid if the cb is null.